pyproject.toml
[tool.poetry]
name = "tonic-validate"
version = "3.2.0"
description = "RAG evaluation metrics."
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.15"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
